+++ pkg/src/cwt_thierry.c
@@ -1,3 +1,5 @@
+#include <stdlib.h>
+
 
 /***************************************************************
 *              (c) Copyright  1997                             *
@@ -24,7 +26,7 @@
 *   M: number of vanishing moments
 ***************************************************************/
 
-void thierry_frequency(int M,float scale,double *w,int isize)
+void thierry_frequency(int M,double scale,double *w,int isize)
 {
   double tmp;
   int i;
@@ -53,11 +55,11 @@ void thierry_frequency(int M,float scale,double *w,int isize)
 *   pcenterfrequency: centralfrequency of Morlet wavelet
 ******************************************************************/
 
-void Scwt_thierry_r(float *input, double *Oreal, double *Oimage,
+void Scwt_thierry_r(double *input, double *Oreal, double *Oimage,
    int *pnboctave, int *pnbvoice, int *pinputsize, int *pM)
 {	
   int nboctave, nbvoice, i, j, inputsize, M;
-  float a;
+  double a;
   double *Ri2, *Ri1, *Ii1, *Ii, *Ri;
 
 
@@ -85,7 +87,7 @@ void Scwt_thierry_r(float *input, double *Oreal, double *Oimage,
   
   for(i = 1; i <= nboctave; i++) {
     for(j=0; j < nbvoice; j++) {
-      a = (float)(pow((double)2,(double)(i+j/((double)nbvoice))));
+      a = (double)(pow((double)2,(double)(i+j/((double)nbvoice))));
       thierry_frequency(M,a,Ri2,inputsize); 
       multi(Ri1,Ii1,Ri2,Oreal,Oimage,inputsize);
       double_fft(Oreal,Oimage,Oreal,Oimage,inputsize,1); 
@@ -120,12 +122,12 @@ void Scwt_thierry_r(float *input, double *Oreal, double *Oimage,
 *   pM: number of vanishing moments
 ******************************************************************/
 
-void Scwt_thierry(float *Rinput,float *Iinput,double *Oreal,
+void Scwt_thierry(double *Rinput,double *Iinput,double *Oreal,
    double *Oimage,int *pnboctave,int *pnbvoice,
    int *pinputsize,int *pM)
 {	
   int nboctave, nbvoice, i, j, k, inputsize, M;
-  float a;
+  double a;
   double *Ri2, *Ri1, *Ii1, *Ii, *Ri;
 
 
@@ -157,7 +159,7 @@ void Scwt_thierry(float *Rinput,float *Iinput,double *Oreal,
   
   for(i = 1; i <= nboctave; i++) {
     for(j=0; j < nbvoice; j++) {
-      a = (float)(pow((double)2,(double)(i+j/((double)nbvoice))));
+      a = (double)(pow((double)2,(double)(i+j/((double)nbvoice))));
       thierry_frequency(M,a,Ri2,inputsize); 
       multi(Ri1,Ii1,Ri2,Oreal,Oimage,inputsize);
       double_fft(Oreal,Oimage,Oreal,Oimage,inputsize,1); 
@@ -187,12 +189,12 @@ void Scwt_thierry(float *Rinput,float *Iinput,double *Oreal,
 *     Oreal, Oimage: real and imaginary parts of the cwt.
 ***************************************************************/
 
-void Svwt_thierry(float *Rinput,float *Iinput,double *Oreal,
-   double *Oimage,float *pa,int *pinputsize,
+void Svwt_thierry(double *Rinput,double *Iinput,double *Oreal,
+   double *Oimage,double *pa,int *pinputsize,
    int *pM)
 {	
   int octave, voice, nbvoice, i, j, k, inputsize, M;
-  float a;
+  double a;
   double *Ri2, *Ri1, *Ii1, *Ii, *Ri;
